@@ -5,28 +5,19 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon -e ts --exec \"npm run compile\"",
-    "bridge:start": "nodemon -e ts --exec \"npm run bridge:compile\"",
-    "bridge:compile": "tsc && node ./build/bridge/index.js",
-    "proposal:start": "nodemon -e ts --exec \"npm run proposal:compile\"",
-    "proposal:compile": "tsc && node ./build/proposal/index.js",
-    "dailyLiquidity:start": "nodemon -e ts --exec \"npm run dailyLiquidity:compile\"",
-    "dailyLiquidity:compile": "tsc && node ./build/dailyLiquidity/index.js",
     "compile": "tsc && node ./build/main.js",
     "build": "tsc",
+    "lint": "eslint ./src --fix",
     "test": "lab --sourcemaps --transform node_modules/lab-transform-typescript -v --leaks"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@pancakeswap/sdk": "^2.4.0",
-    "@uniswap/sdk": "^3.0.3",
-    "@uniswap/v2-core": "^1.0.1",
     "@ethersproject/address": "^5.4.0",
     "@ethersproject/contracts": "^5.4.1",
     "@ethersproject/providers": "^5.4.5",
     "@ethersproject/solidity": "^5.4.0",
-    "web3": "^1.5.3",
     "@hapi/basic": "^5.1.1",
     "@hapi/bell": "^12.2.0",
     "@hapi/boom": "^9.1.0",
@@ -39,17 +30,18 @@
     "@types/dotenv": "^6.1.1",
     "@types/joi": "^14.3.4",
     "@types/validator": "^13.1.0",
-    "@workquest/database-models": "git+https://github.com/WorkQuest/database-models.git#ba6a59a89040c3ae265b0648cb6b46a25274e226",
+    "@uniswap/sdk": "^3.0.3",
+    "@uniswap/v2-core": "^1.0.1",
+    "@workquest/database-models": "git+https://github.com/WorkQuest/database-models.git#995db3c561ff812414a0632b5bab9ea3defbbe15",
+    "amqplib": "^0.8.0",
     "aws-sdk": "^2.918.0",
     "axios": "^0.21.1",
     "bcrypt": "^5.0.0",
     "bech32-converting": "^1.0.5",
     "bignumber.js": "^9.0.1",
     "dotenv": "^8.2.0",
-    "ethereum-block-by-date": "^1.4.2",
     "ethers": "^5.5.1",
     "filepaths": "^0.3.0",
-    "form-data": "^4.0.0",
     "geoip-lite": "^1.4.2",
     "graphile-worker": "^0.6.0",
     "handlebars": "^4.7.7",
@@ -61,11 +53,9 @@
     "joi": "^17.4.0",
     "jsonwebtoken": "^8.5.1",
     "lab-transform-typescript": "^3.0.1",
-    "node-cron": "^3.0.0",
     "node-fetch": "^3.0.0",
     "nodemailer": "^6.5.0",
     "nodemon": "^2.0.12",
-    "npm": "^7.14.0",
     "pg": "^8.3.0",
     "qs": "^6.9.6",
     "reflect-metadata": "^0.1.13",
@@ -74,26 +64,19 @@
     "speakeasy": "^2.0.0",
     "twilio": "^3.66.0",
     "typescript": "^4.1.5",
-    "uuid": "^8.2.0"
+    "uuid": "^8.2.0",
+    "web3": "^1.5.3"
   },
   "devDependencies": {
     "@hapi/code": "^8.0.3",
-    "@hapi/eslint-config-hapi": "^12.3.0",
-    "@hapi/eslint-plugin-hapi": "^4.3.4",
     "@hapi/lab": "^24.2.1",
     "@types/hapi": "^18.0.3",
     "@types/node": "^12.12.52",
-    "eslint": "^6.5.1",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.4.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-standard": "^4.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.8.0",
+    "@typescript-eslint/parser": "^5.8.0",
+    "eslint": "^8.7.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "pino": "^6.13.2"
   }
 }
