@@ -1,59 +1,90 @@
-import { literal, Op } from "sequelize";
-import { Errors } from "../utils/errors";
-import { UserController } from "../controllers/user/controller.user";
-import { QuestController } from "../controllers/quest/controller.quest";
-import { transformToGeoPostGIS } from "../utils/postGIS";
-import { error, output } from "../utils";
-import { publishQuestNotifications, QuestNotificationActions } from "../websocket/websocket.quest";
-import { QuestsResponseController } from "../controllers/quest/controller.questsResponse";
-import { MediaController } from "../controllers/controller.media";
-import { SkillsFiltersController } from "../controllers/controller.skillsFilters";
-import { addUpdateReviewStatisticsJob } from "../jobs/updateReviewStatistics";
-import { updateQuestsStatisticJob } from "../jobs/updateQuestsStatistic";
+import { literal, Op } from 'sequelize';
+import { Errors } from '../utils/errors';
+import { UserController } from '../controllers/user/controller.user';
+import { QuestController } from '../controllers/quest/controller.quest';
+import { transformToGeoPostGIS } from '../utils/postGIS';
+import { error, output } from '../utils';
+import { QuestNotificationActions } from '../controllers/controller.broker';
+import { QuestsResponseController } from '../controllers/quest/controller.questsResponse';
+import { MediaController } from '../controllers/controller.media';
+import { addUpdateReviewStatisticsJob } from '../jobs/updateReviewStatistics';
+import { updateQuestsStatisticJob } from '../jobs/updateQuestsStatistic';
+import { SkillsFiltersController } from '../controllers/controller.skillsFilters';
 import {
-  Chat,
-  User,
+  DisputeStatus,
   Quest,
-  Review,
-  UserRole,
   QuestChat,
-  QuestStatus,
-  StarredQuests,
-  QuestsResponse,
   QuestChatStatuses,
+  QuestDispute,
+  QuestsResponse,
+  QuestsResponseStatus,
   QuestsResponseType,
-  QuestSpecializationFilter,
+  QuestStatus,
+  Review,
+  StarredQuests,
+  User,
+  UserRole
 } from "@workquest/database-models/lib/models";
 
-export const searchFields = [
-  "title",
-  "description",
+export const searchQuestFields = [
+  'title',
+  'description',
+  'locationPlaceName'
 ];
 
 export async function getQuest(r) {
   const user: User = r.auth.credentials;
 
-  const quest = await Quest.findOne({
-    where: { id: r.params.questId },
-    include: [{
-      model: StarredQuests,
-      as: "star",
-      where: { userId: r.auth.credentials.id },
-      required: false
-    }, {
-      model: QuestsResponse,
-      as: "response",
-      where: { workerId: r.auth.credentials.id },
-      required: false
-    }, {
+  const include = [{
+    model: StarredQuests,
+    as: "star",
+    where: { userId: r.auth.credentials.id },
+    required: false
+  }, {
+    model: QuestsResponse,
+    as: "response",
+    where: { workerId: r.auth.credentials.id },
+    required: false
+  }, {
+    model: User.scope('shortWithWallet'),
+    as: 'user'
+  }, {
+    model: User.scope('shortWithWallet'),
+    as: 'assignedWorker'
+  }, {
+    model: QuestDispute.unscoped(),
+    as: 'openDispute',
+    where: {
+      [Op.or]: [
+        { opponentUserId: r.auth.credentials.id },
+        { openDisputeUserId: r.auth.credentials.id },
+      ],
+      status: { [Op.in]: [DisputeStatus.pending, DisputeStatus.inProgress] },
+    },
+    required: false,
+  }, {
+    model: Review.unscoped(),
+    as: 'yourReview',
+    where: { fromUserId: r.auth.credentials.id },
+    required: false,
+  }] as any[];
+
+  if (user.role === UserRole.Worker) {
+    include.push({
       model: QuestChat.scope('idsOnly'),
       as: 'questChat',
       required: false,
-    }]
+      where: { workerId: user.id },
+    });
+  }
+
+  const quest = await Quest.findOne({
+    where: { id: r.params.questId },
+    include,
   });
 
   if (!quest) {
-    return error(Errors.NotFound, "Quest not found", { questId: r.params.questId });
+    return error(Errors.NotFound, 'Quest not found', { questId: r.params.questId });
   }
 
   return output(quest);
@@ -76,14 +107,14 @@ export async function createQuest(r) {
     workplace: r.payload.workplace,
     employment: r.payload.employment,
     priority: r.payload.priority,
-    location: r.payload.location,
     title: r.payload.title,
     description: r.payload.description,
     price: r.payload.price,
     medias: r.payload.medias,
     adType: r.payload.adType,
-    locationPlaceName: r.payload.locationPlaceName,
-    locationPostGIS: transformToGeoPostGIS(r.payload.location),
+    location: r.payload.locationFull.location,
+    locationPlaceName: r.payload.locationFull.locationPlaceName,
+    locationPostGIS: transformToGeoPostGIS(r.payload.locationFull.location),
   }, { transaction });
 
   const questController = new QuestController(quest);
@@ -98,9 +129,7 @@ export async function createQuest(r) {
     role: UserRole.Employer,
   });
 
-  return output(
-    await Quest.findByPk(quest.id)
-  )
+  return output(await Quest.findByPk(quest.id));
 }
 
 export async function editQuest(r) {
@@ -128,12 +157,24 @@ export async function editQuest(r) {
     workplace: r.payload.workplace,
     employment: r.payload.employment,
     description: r.payload.description,
-    location: r.payload.location,
-    locationPlaceName: r.payload.locationPlaceName,
-    locationPostGIS: transformToGeoPostGIS(r.payload.location),
+    location: r.payload.locationFull.location,
+    locationPlaceName: r.payload.locationFull.locationPlaceName,
+    locationPostGIS: transformToGeoPostGIS(r.payload.locationFull.location),
   }, { transaction });
 
   await transaction.commit();
+
+  const responses = await QuestsResponse.findAll({
+    where: { questId: questController.quest.id, status: QuestsResponseStatus.Open },
+  });
+
+  if (responses.length !== 0) {
+    r.server.app.broker.sendQuestNotification({
+      action: QuestNotificationActions.questEdited,
+      recipients: responses.map(_ => _.workerId),
+      data: questController.quest,
+    });
+  }
 
   return output(questController.quest);
 }
@@ -209,10 +250,10 @@ export async function startQuest(r) {
 
   await transaction.commit();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: questController.quest,
     recipients: [assignedWorkerController.user.id],
-    action: QuestNotificationActions.questStarted,
+    action: QuestNotificationActions.waitWorker,
   });
 
   return output();
@@ -238,7 +279,7 @@ export async function rejectWorkOnQuest(r) {
 
   await transaction.commit();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     recipients: [questController.quest.userId],
     action: QuestNotificationActions.workerRejectedQuest,
     data: questController.quest,
@@ -272,7 +313,7 @@ export async function acceptWorkOnQuest(r) {
     role: UserRole.Worker,
   });
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: questController.quest,
     recipients: [questController.quest.userId],
     action: QuestNotificationActions.workerAcceptedQuest,
@@ -293,7 +334,7 @@ export async function completeWorkOnQuest(r) {
 
   await questController.completeWork();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: questController.quest,
     recipients: [questController.quest.userId],
     action: QuestNotificationActions.workerCompletedQuest,
@@ -314,7 +355,7 @@ export async function acceptCompletedWorkOnQuest(r) {
 
   await questController.approveCompletedWork();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: quest,
     recipients: [quest.assignedWorkerId],
     action: QuestNotificationActions.employerAcceptedCompletedQuest,
@@ -351,7 +392,7 @@ export async function rejectCompletedWorkOnQuest(r) {
 
   await questController.rejectCompletedWork();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: quest,
     recipients: [quest.assignedWorkerId],
     action: QuestNotificationActions.employerRejectedCompletedQuest,
@@ -361,92 +402,141 @@ export async function rejectCompletedWorkOnQuest(r) {
 }
 
 export async function getQuests(r) {
+  const user: User = r.auth.credentials;
+
   const entersAreaLiteral = literal(
     'st_within("Quest"."locationPostGIS", st_makeenvelope(:northLng, :northLat, :southLng, :southLat, 4326))'
   );
+  const questChatLiteral = literal(
+    'CASE WHEN "questChat->quest" = NULL THEN NULL ELSE "questChat->quest"."id" END'
+  );
+  const questSpecializationOnlyPathsLiteral = literal(
+    '(1 = (CASE WHEN EXISTS (SELECT "id" FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."path" IN (:path)) THEN 1 END))'
+  );
+  const questSpecializationOnlyIndustryKeysLiteral = literal(
+    '(1 = (CASE WHEN EXISTS (SELECT * FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."industryKey" IN (:industryKey)) THEN 1 END))'
+  );
+  const questSpecializationIndustryKeysAndPathsLiteral = literal(
+    '(1 = (CASE WHEN EXISTS (SELECT * FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."path" IN (:path)) THEN 1 END))' +
+    'OR (1 = (CASE WHEN EXISTS (SELECT * FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."industryKey" IN (:industryKey)) THEN 1 END))'
+  );
+  const userSearchLiteral = literal(
+    // TODO добавь эти поля в replace типо так ILIKE '%:searchByFirstName%'`
+    `(SELECT "firstName" FROM "Users" WHERE "id" = "Quest"."userId") ILIKE '%${r.query.q}%'` +
+    `OR (SELECT "lastName" FROM "Users" WHERE "id" = "Quest"."userId") ILIKE '%${r.query.q}%'`
+  );
+  const questChatWorkerLiteral = literal(
+    '"questChat"."workerId" = "Quest"."assignedWorkerId"'
+  );
+
   const order = [];
   const include = [];
+  const replacements = {};
   const where = {
-    ...(r.query.statuses && { status: { [Op.in]: r.query.statuses } }),
+    [Op.and]: [],
     ...(r.query.adType && { adType: r.query.adType }),
     ...(r.query.filter && { filter: r.params.filter }),
     ...(r.params.userId && { userId: r.params.userId }),
     ...(r.params.workerId && { assignedWorkerId: r.params.workerId }),
+    ...(r.query.statuses && { status: { [Op.in]: r.query.statuses } }),
     ...(r.query.performing && { assignedWorkerId: r.auth.credentials.id }),
-    ...(r.query.north && r.query.south && { [Op.and]: entersAreaLiteral }),
-    ...(r.query.priorities && { priority: {[Op.in]: r.query.priorities } }),
+    ...(r.query.priorities && { priority: { [Op.in]: r.query.priorities } }),
     ...(r.query.workplaces && { workplace: { [Op.in]: r.query.workplaces } }),
     ...(r.query.employments && { employment: { [Op.in]: r.query.employments } }),
     ...(r.query.priceBetween && { price: { [Op.between]: [r.query.priceBetween.from, r.query.priceBetween.to] } }),
   };
 
   if (r.query.q) {
-    where[Op.or] = searchFields.map(field => ({
+    where[Op.or] = searchQuestFields.map(field => ({
       [field]: { [Op.iLike]: `%${r.query.q}%` }
     }));
-  }
-  if (r.query.specializations) {
-    const { specializationKeys, industryKeys } = SkillsFiltersController.splitSpecialisationAndIndustry(r.query.specializations);
 
+    where[Op.or].push(userSearchLiteral)
+  }
+  if (r.query.specializations) { // TODO r.query.specialization on r.query.specialization[s]
+    const { paths, industryKeys } = SkillsFiltersController.splitPathsAndSingleKeysOfIndustry(r.query.specializations);
+
+    if (paths.length !== 0 && industryKeys.length === 0) {
+      replacements['path'] = paths;
+      where[Op.and].push(questSpecializationOnlyPathsLiteral);
+    }
+    if (paths.length === 0 && industryKeys.length !== 0) {
+      replacements['industryKey'] = industryKeys;
+      where[Op.and].push(questSpecializationOnlyIndustryKeysLiteral);
+    }
+    if (paths.length !== 0 && industryKeys.length !== 0) {
+      replacements['path'] = paths;
+      replacements['industryKey'] = industryKeys;
+      where[Op.and].push(questSpecializationIndustryKeysAndPathsLiteral);
+    }
+  }
+  if (r.query.northAndSouthCoordinates) {
+    replacements['northLng'] = r.query.northAndSouthCoordinates.north.longitude;
+    replacements['northLat'] = r.query.northAndSouthCoordinates.north.latitude;
+    replacements['southLng'] = r.query.northAndSouthCoordinates.south.longitude;
+    replacements['southLat'] = r.query.northAndSouthCoordinates.south.latitude;
+
+    where[Op.and].push(entersAreaLiteral);
+  }
+  if (user.role === UserRole.Worker) {
     include.push({
-      model: QuestSpecializationFilter,
-      as: 'questIndustryForFiltering',
-      attributes: [],
-      where: { industryKey: { [Op.in]: industryKeys } }
+      model: QuestChat.scope('idsOnly'),
+      where: { workerId: user.id },
+      as: 'questChat',
+      required: false,
     });
-
-    if (specializationKeys.length > 0) {
-      include.push({
-        model: QuestSpecializationFilter,
-        as: 'questSpecializationForFiltering',
-        attributes: [],
-        where: { specializationKey: { [Op.in]: specializationKeys } }
-      });
-    }
+  }
+  if (user.role === UserRole.Employer) {
+    include.push({
+      model: QuestChat.scope('idsOnly'),
+      as: 'questChat',
+      attributes: {
+        include: [[questChatLiteral, 'id']],
+      },
+      where: { employerId: user.id, questChatWorkerLiteral },
+      required: false,
+      include: {
+        model: Quest.unscoped(),
+        as: 'quest',
+        attributes: ['id', 'status'],
+        where: {
+          status: [QuestStatus.Active, QuestStatus.Dispute, QuestStatus.WaitWorker, QuestStatus.WaitConfirm],
+        },
+        required: false,
+      },
+    });
   }
 
-  include.push({
-    model: Review.unscoped(),
-    as: "yourReview",
-    where: { fromUserId: r.auth.credentials.id },
-    required: false,
-  }, {
-    model: StarredQuests.unscoped(),
-    as: "star",
-    where: { userId: r.auth.credentials.id },
-    required: r.query.starred,
-  }, {
-    model: QuestsResponse.unscoped(),
-    as: 'invited',
-    required: r.query.invited,
-    where: {
-      [Op.and]: [
-        { workerId: r.auth.credentials.id },
-        { type: QuestsResponseType.Invite },
-      ]
-    }
-  }, {
-    model: QuestsResponse.unscoped(),
-    as: "responded",
-    required: r.query.responded,
-    where: {
-      [Op.and]: [
-        { workerId: r.auth.credentials.id },
-        { type: QuestsResponseType.Response },
-      ]
+  include.push(
+    {
+      model: Review.unscoped(),
+      as: 'yourReview',
+      where: { fromUserId: r.auth.credentials.id },
+      required: false,
     },
-  }, {
-    model: QuestChat.scope('idsOnly'),
-    as: 'questChat',
-    required: false,
-  });
-
-  // {
-  //   model: QuestsResponse,
-  //     as: 'responses',
-  //   required: false,
-  //   where: { '$"Quest"."userId"$': r.auth.credentials.id },
-  // }
+    {
+      model: StarredQuests.unscoped(),
+      as: 'star',
+      where: { userId: r.auth.credentials.id },
+      required: r.query.starred,
+    },
+    {
+      model: QuestsResponse.unscoped(),
+      as: 'invited',
+      required: r.query.invited,
+      where: {
+        [Op.and]: [{ workerId: r.auth.credentials.id }, { type: QuestsResponseType.Invite }],
+      },
+    },
+    {
+      model: QuestsResponse.unscoped(),
+      as: 'responded',
+      required: r.query.responded,
+      where: {
+        [Op.and]: [{ workerId: r.auth.credentials.id }, { type: QuestsResponseType.Response }],
+      },
+    },
+  );
 
   for (const [key, value] of Object.entries(r.query.sort)) {
     order.push([key, value]);
@@ -456,15 +546,10 @@ export async function getQuests(r) {
     distinct: true,
     limit: r.query.limit,
     offset: r.query.offset,
-    include, order, where,
-    replacements: {
-      ...(r.query.north && r.query.south && {
-        northLng: r.query.north.longitude,
-        northLat: r.query.north.latitude,
-        southLng: r.query.south.longitude,
-        southLat: r.query.south.latitude,
-      })
-    }
+    include,
+    order,
+    where,
+    replacements,
   });
 
   return output({ count, quests: rows });
