@@ -1,58 +1,23 @@
-import { Op, Transaction } from "sequelize";
-import {error} from "../../utils";
-import {Errors} from '../../utils/errors';
-import config from "../../config/config";
-import {totpValidate} from "@workquest/database-models/lib/utils";
-import {SkillsFiltersController} from "../controller.skillsFilters";
+import { Op, Transaction } from 'sequelize';
+import { error } from '../../utils';
+import { Errors } from '../../utils/errors';
+import config from '../../config/config';
+import { totpValidate } from '@workquest/database-models/lib/utils';
+import { SkillsFiltersController } from '../controller.skillsFilters';
 import {
   User,
+  Session,
   UserRole,
   UserStatus,
+  ChatsStatistic,
+  QuestsStatistic,
   RatingStatistic,
   defaultUserSettings,
-  UserSpecializationFilter, Session
-} from "@workquest/database-models/lib/models";
+  UserSpecializationFilter,
+} from '@workquest/database-models/lib/models';
 
 abstract class UserHelper {
-  public abstract user: User
-
-  public async setUserSpecializations(keys: string[], transaction?: Transaction) {
-    try {
-      await UserSpecializationFilter.destroy({
-        where: { userId: this.user.id }, transaction,
-      });
-
-      if (keys.length <= 0) {
-        return;
-      }
-
-      const skillsFiltersController = await SkillsFiltersController.getInstance();
-      const userSpecializations = skillsFiltersController.keysToRecords(keys, 'userId', this.user.id);
-
-      await UserSpecializationFilter.bulkCreate(userSpecializations, { transaction });
-    } catch (e) {
-      if (transaction) {
-        await transaction.rollback();
-        throw e;
-      }
-    }
-  }
-
-  public async logoutAllSessions(transaction?: Transaction) {
-    try {
-      await Session.update({ invalidating: true }, {
-        where: {
-          userId: this.user.id,
-          createdAt: { [Op.gte]: Date.now() - config.auth.jwt.refresh.lifetime * 1000 },
-        }, transaction,
-      });
-    } catch (e) {
-      if (transaction) {
-        await transaction.rollback();
-      }
-      throw e;
-    }
-  }
+  public abstract user: User;
 
   public static getDefaultAdditionalInfo(role: UserRole) {
     let additionalInfo: object = {
@@ -63,8 +28,8 @@ abstract class UserHelper {
         instagram: null,
         twitter: null,
         linkedin: null,
-        facebook: null
-      }
+        facebook: null,
+      },
     };
 
     if (role === UserRole.Worker) {
@@ -72,14 +37,14 @@ abstract class UserHelper {
         ...additionalInfo,
         skills: [],
         educations: [],
-        workExperiences: []
+        workExperiences: [],
       };
     } else if (role === UserRole.Employer) {
       additionalInfo = {
         ...additionalInfo,
         company: null,
         CEO: null,
-        website: null
+        website: null,
       };
     }
 
@@ -115,11 +80,11 @@ abstract class UserHelper {
       status: UserStatus.NeedSetRole,
       email: profile.email.toLowerCase(),
       settings: Object.assign({}, defaultUserSettings, {
-        social: { [network]: socialInfo }
-      })
+        social: { [network]: socialInfo },
+      }),
     });
 
-    await RatingStatistic.create({ userId: user.id });
+    await UserController.createStatistics(user.id);
 
     return user;
   }
@@ -134,20 +99,21 @@ abstract class UserHelper {
   }
 
   public static async userMustExist(userId: string) {
-    if (!await User.findByPk(userId)) {
-      throw error(Errors.NotFound, "User does not exist", { userId });
+    if (!(await User.findByPk(userId))) {
+      throw error(Errors.NotFound, 'User does not exist', { userId });
     }
   }
 
-  public static async usersMustExist(userIds: string[], scope: "defaultScope" | "short" | "shortWithAdditionalInfo" = "defaultScope"): Promise<User[]> {
-   const users = await User.scope(scope).findAll({
-      where: { id: userIds }
-   });
+  public static async usersMustExist(
+    userIds: string[],
+    scope: 'defaultScope' | 'short' | 'shortWithAdditionalInfo' = 'defaultScope',
+  ): Promise<User[]> {
+    const users = await User.scope(scope).findAll({
+      where: { id: userIds },
+    });
 
     if (users.length !== userIds.length) {
-      const notFoundIds = userIds.filter(id =>
-        users.findIndex(user => id === user.id) === -1
-      );
+      const notFoundIds = userIds.filter((userId) => users.findIndex((user) => userId === user.id) === -1);
 
       throw error(Errors.NotFound, 'Users is not found', { notFoundIds });
     }
@@ -159,14 +125,26 @@ abstract class UserHelper {
     const emailUsed = await User.findOne({ where: { email: { [Op.iLike]: email } } });
 
     if (emailUsed) {
-      throw error(Errors.InvalidPayload, "Email used", [{ field: "email", reason: "used" }]);
+      throw error(Errors.InvalidPayload, 'Email used', [{ field: 'email', reason: 'used' }]);
     }
   }
 
   public userMustHaveRole(role: UserRole): this {
     if (this.user.role !== role) {
       throw error(Errors.InvalidRole, "User isn't match role", {
-        current: this.user.role, mustHave: role,
+        current: this.user.role,
+        mustHave: role,
+      });
+    }
+
+    return this;
+  }
+
+  public userMustHaveStatus(...statuses: UserStatus[]): this {
+    if (!statuses.includes(this.user.status)) {
+      throw error(Errors.InvalidStatus, "User status doesn't match", {
+        current: this.user.status,
+        mustHave: statuses,
       });
     }
 
@@ -201,8 +179,7 @@ abstract class UserHelper {
 
   public userMustHaveActiveStatusTOTP(activeStatus: boolean): this {
     if (this.user.settings.security.TOTP.active !== activeStatus) {
-      throw error(Errors.InvalidActiveStatusTOTP,
-        `Active status TOTP is not ${activeStatus ? "enable" : "disable"}`, {});
+      throw error(Errors.InvalidActiveStatusTOTP, `Active status TOTP is not ${activeStatus ? 'enable' : 'disable'}`, {});
     }
 
     return this;
@@ -218,7 +195,7 @@ abstract class UserHelper {
 
   public checkTotpConfirmationCode(code): this {
     if (!totpValidate(code, this.user.settings.security.TOTP.secret)) {
-      throw error(Errors.InvalidPayload, "TOTP is invalid", [{ field: "totp", reason: "invalid" }]);
+      throw error(Errors.InvalidPayload, 'TOTP is invalid', [{ field: 'totp', reason: 'invalid' }]);
     }
 
     return this;
@@ -226,10 +203,12 @@ abstract class UserHelper {
 
   public checkActivationCodeTotp(code): this {
     if (this.user.settings.security.TOTP.confirmCode !== code) {
-      throw error(Errors.InvalidPayload, "Confirmation code is not correct", [{
-        field: "confirmCode",
-        reason: "invalid"
-      }]);
+      throw error(Errors.InvalidPayload, 'Confirmation code is not correct', [
+        {
+          field: 'confirmCode',
+          reason: 'invalid',
+        },
+      ]);
     }
 
     return this;
@@ -237,7 +216,7 @@ abstract class UserHelper {
 
   public checkUserAlreadyConfirmed(): this {
     if (!this.user.settings.emailConfirm) {
-      throw error(Errors.UserAlreadyConfirmed, "User already confirmed", {});
+      throw error(Errors.UserAlreadyConfirmed, 'User already confirmed', {});
     }
 
     return this;
@@ -245,7 +224,7 @@ abstract class UserHelper {
 
   public checkUserConfirmationCode(confirmCode): this {
     if (this.user.settings.emailConfirm.toLowerCase() !== confirmCode.toLowerCase()) {
-      throw error(Errors.InvalidPayload, "Invalid confirmation code", [{ field: "confirmCode", reason: "invalid" }]);
+      throw error(Errors.InvalidPayload, 'Invalid confirmation code', [{ field: 'confirmCode', reason: 'invalid' }]);
     }
 
     return this;
@@ -253,21 +232,19 @@ abstract class UserHelper {
 }
 
 export class UserController extends UserHelper {
-
-  constructor(
-    public user: User
-  ) {
+  constructor(public user: User) {
     super();
 
     if (!user) {
-      throw error(Errors.NotFound, "User not found", {});
+      throw error(Errors.NotFound, 'User not found', {});
     }
   }
 
   public async setRole(role: UserRole, transaction?: Transaction) {
     try {
       this.user = await this.user.update({
-        status: UserStatus.Confirmed, role,
+        status: UserStatus.Confirmed,
+        role,
         additionalInfo: UserController.getDefaultAdditionalInfo(role),
       });
     } catch (e) {
@@ -278,12 +255,14 @@ export class UserController extends UserHelper {
     }
   }
 
-  public async setUnverifiedPhoneNumber(phoneNumber: string, confirmCode: number, transaction?: Transaction) {
+  public async setConfirmCodeToVerifyCodeNumber(confirmCode: number, transaction?: Transaction) {
     try {
-      await this.user.update({
-        tempPhone: phoneNumber,
-        'settings.phoneConfirm': confirmCode,
-      }, { transaction });
+      await this.user.update(
+        {
+          'settings.phoneConfirm': confirmCode,
+        },
+        { transaction },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -298,7 +277,7 @@ export class UserController extends UserHelper {
         phone: this.user.tempPhone,
         tempPhone: null,
         'settings.phoneConfirm': null,
-      });
+      }, { transaction });
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -307,16 +286,88 @@ export class UserController extends UserHelper {
     }
   }
 
-  public async changePassword(newPassword: String, transaction?: Transaction) {
+  public async changePassword(newPassword: string, transaction?: Transaction) {
     try {
-      this.user = await this.user.update({
-        password: newPassword
-      }, { transaction });
+      this.user = await this.user.update(
+        {
+          password: newPassword,
+        },
+        { transaction },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
       }
       throw e;
     }
+  }
+
+  public async logoutAllSessions(transaction?: Transaction) {
+    try {
+      await Session.update(
+        { invalidating: true },
+        {
+          where: {
+            userId: this.user.id,
+            createdAt: { [Op.gte]: Date.now() - config.auth.jwt.refresh.lifetime * 1000 },
+          },
+          transaction,
+        },
+      );
+    } catch (e) {
+      if (transaction) {
+        await transaction.rollback();
+      }
+      throw e;
+    }
+  }
+
+  public async setUserSpecializations(keys: string[], transaction?: Transaction) {
+    try {
+      await UserSpecializationFilter.destroy({
+        where: { userId: this.user.id },
+        transaction,
+      });
+
+      if (keys.length <= 0) {
+        return;
+      }
+
+      const skillsFiltersController = await SkillsFiltersController.getInstance();
+      const userSpecializations = skillsFiltersController.keysToRecords(keys, 'userId', this.user.id);
+
+      await UserSpecializationFilter.bulkCreate(userSpecializations, { transaction });
+    } catch (e) {
+      if (transaction) {
+        await transaction.rollback();
+        throw e;
+      }
+    }
+  }
+
+  public static async createStatistics(userId) {
+    await RatingStatistic.findOrCreate({
+      where: { userId: userId },
+      defaults: { userId: userId },
+    });
+    await ChatsStatistic.findOrCreate({
+      where: { userId: userId },
+      defaults: { userId: userId },
+    });
+    await QuestsStatistic.findOrCreate({
+      where: { userId: userId },
+      defaults: { userId: userId },
+    });
+  }
+
+  public get shortCredentials() {
+    return {
+      id: this.user.id,
+      firstName: this.user.firstName,
+      lastName: this.user.lastName,
+      avatarId: this.user.avatarId,
+      avatar: this.user.avatar,
+      additionalInfo: this.user.additionalInfo,
+    };
   }
 }
