@@ -13,7 +13,10 @@ export enum Errors {
   UserAlreadyConfirmed = 400008,
   InvalidType = 400009,
   AlreadyExists = 400010,
-  WalletExists = 400011,
+  InvalidDate = 400011,
+  WalletExists = 400012,
+  BlockedUser = 400013,
+  PhoneNumberAlreadyConfirmed = 400014,
   // Authorization errors (401)
   TokenExpired = 401001,
   TokenInvalid = 401002,
