@@ -1,7 +1,7 @@
-import { v4 as uuidv4 } from "uuid";
-import { Boom } from "@hapi/boom";
-import * as crypto from "crypto";
-
+import { v4 as uuidv4 } from 'uuid';
+import { Boom } from '@hapi/boom';
+import * as crypto from 'crypto';
+import config from '../config/config';
 const geoip = require('geoip-lite');
 
 export function getUUID(): string {
@@ -9,32 +9,38 @@ export function getUUID(): string {
 }
 
 export function getRealIp(request): string {
-  return request.headers["cf-connecting-ip"] ?
-    request.headers["cf-connecting-ip"] :
-    request.info.remoteAddress;
+  return request.headers['X-Forwarded-For'] ? request.headers['X-Forwarded-For'] : request.info.remoteAddress;
 }
 
-export function getGeo(request): { country: string, city: string } {
+//DO NOT WORK WITH LOCAL IP
+export function getGeo(request) {
+  if (config.debug) {
+    return {
+      country: 'localhost',
+      city: 'localhost',
+    };
+  }
+
   const ip = getRealIp(request);
   const geo = geoip.lookup(ip);
 
   return {
-    country: geo ? geo.country : null,
-    city: geo ? geo.city : null,
-  }
+    country: geo.country,
+    city: geo.city,
+  };
 }
 
 export function getDevice(request): string {
   return request.headers['user-agent'];
 }
 
-export function output(result?: any | null): { ok: boolean, result: any | null } {
+export function output(result?: any | null): { ok: boolean; result: any | null } {
   return { ok: true, result };
 }
 
 export interface OutputInterface {
-  ok: boolean,
-  result?: object
+  ok: boolean;
+  result?: object;
 }
 
 export function error(code: number, msg: string, data: object): Boom {
@@ -42,7 +48,7 @@ export function error(code: number, msg: string, data: object): Boom {
     data: {
       code,
       data,
-      api: true
+      api: true,
     },
     statusCode: Math.floor(code / 1000),
   });
@@ -51,34 +57,40 @@ export function error(code: number, msg: string, data: object): Boom {
 export function responseHandler(r, h) {
   // Handle default hapi errors (like not found, etc.)
   if (r.response.isBoom && r.response.data === null) {
-    r.response = h.response({
-      ok: false,
-      code: Math.floor(r.response.output.statusCode * 1000),
-      data: {},
-      msg: r.response.message
-    }).code(r.response.output.statusCode);
+    r.response = h
+      .response({
+        ok: false,
+        code: Math.floor(r.response.output.statusCode * 1000),
+        data: {},
+        msg: r.response.message,
+      })
+      .code(r.response.output.statusCode);
 
     return h.continue;
   }
   // Handle custom api error
   if (r.response.isBoom && r.response.data.api) {
-    r.response = h.response({
-      ok: false,
-      code: r.response.data.code,
-      data: r.response.data.data,
-      msg: r.response.output.payload.message
-    }).code(Math.floor(r.response.data.code / 1000));
+    r.response = h
+      .response({
+        ok: false,
+        code: r.response.data.code,
+        data: r.response.data.data,
+        msg: r.response.output.payload.message,
+      })
+      .code(Math.floor(r.response.data.code / 1000));
 
     return h.continue;
   }
   // Handle non api errors with data
   if (r.response.isBoom && !r.response.data.api) {
-    r.response = h.response({
-      ok: false,
-      code: Math.floor(r.response.output.statusCode * 1000),
-      data: r.response.data,
-      msg: r.response.message
-    }).code(r.response.output.statusCode);
+    r.response = h
+      .response({
+        ok: false,
+        code: Math.floor(r.response.output.statusCode * 1000),
+        data: r.response.data,
+        msg: r.response.message,
+      })
+      .code(r.response.output.statusCode);
 
     return h.continue;
   }
@@ -87,7 +99,7 @@ export function responseHandler(r, h) {
 }
 
 export function getRandomHexToken(): string {
-  return crypto.randomBytes(20).toString("hex");
+  return crypto.randomBytes(20).toString('hex');
 }
 
 export function getRandomCodeNumber(): number {
@@ -95,8 +107,11 @@ export function getRandomCodeNumber(): number {
 }
 
 export async function handleValidationError(r, h, err) {
-  return error(400000, "Validation error", err.details.map(e => {
-      return { field: e.context.key, reason: e.type.replace("any.", "") };
-    })
+  return error(
+    400000,
+    'Validation error',
+    err.details.map((e) => {
+      return { field: e.context.key, reason: e.type.replace('any.', '') };
+    }),
   );
 }
